+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+`1.5`_ (2019-01-12)
+~~~~~~~~~~~~~~~~~~~
+
 - Changed the meta tags dictionary to also build absolute URIs when
   calling ``update()`` after the ``meta_tags`` helper returns.
 - Changed ``meta_tags`` to no longer produce a ``<meta
@@ -67,4 +70,5 @@ Change log
 .. _1.2: https://github.com/matthiask/feincms3-meta/compare/1.1...1.2
 .. _1.3: https://github.com/matthiask/feincms3-meta/compare/1.2...1.3
 .. _1.4: https://github.com/matthiask/feincms3-meta/compare/1.3...1.4
-.. _Next version: https://github.com/matthiask/feincms3-meta/compare/1.4...master
+.. _1.5: https://github.com/matthiask/feincms3-meta/compare/1.4...1.5
+.. _Next version: https://github.com/matthiask/feincms3-meta/compare/1.5...master